Also add _add_header() to crl_client()
