+++ certvalidator/ocsp_client.py
@@ -9,10 +9,10 @@ from asn1crypto import core, ocsp, x509, algos, util
 from . import errors
 from ._types import str_cls, type_name
 from .version import __version__
+from ._urllib import _get_host, _add_header
 
 if sys.version_info < (3,):
     from urllib2 import Request, urlopen, URLError
-
 else:
     from urllib.request import Request, urlopen
     from urllib.error import URLError
@@ -117,41 +117,3 @@ def fetch(cert, issuer, hash_algo='sha1', nonce=True, user_agent=None, timeout=1
             last_e = e
 
     raise last_e
-
-
-def _add_header(request, name, value):
-    """
-    Adds a header to a urllib2/urllib.request Request object, ensuring values
-    are encoded appropriately based on the version of Python
-
-    :param request:
-        An instance of urllib2.Request or urllib.request.Request
-
-    :param name:
-        A unicode string of the header name
-
-    :param value:
-        A unicode string of the header value
-    """
-
-    if sys.version_info < (3,):
-        name = name.encode('iso-8859-1')
-        value = value.encode('iso-8859-1')
-
-    request.add_header(name, value)
-
-
-def _get_host(request):
-    """
-    Get's the hostname from the request object according to the python version.
-
-    :param request:
-        An instance of urllib2.Request or urllib.request.Request
-
-    :returns:
-        A string containing the hostname without colon and portnumber.
-    """
-
-    if sys.version_info < (3,):
-        return request.get_host().split(":")[0]
-    return request.host.split(":")[0]
